+++ NEWS.md
@@ -15,6 +15,8 @@ shiny 1.4.0.9001
 
 * Resolved [#2732](https://github.com/rstudio/shiny/issues/2732): `markdown()` is a new function for writing Markdown with Github extensions directly in Shiny UIs. Markdown rendering is performed by the [commonmark](https://github.com/jeroen/commonmark) package. ([#2737](https://github.com/rstudio/shiny/pull/2737))
 
+* `tabsetPanel()` added a new option `type = "hidden"` which hides the tab titles. When using this option, `updateTabsetPanel()` must be called to switch tabs. `type = "hidden"` allows for quick switching between different UI sets.  ([#2814](https://github.com/rstudio/shiny/pull/2814))
+
 ### Minor new features and improvements
 
 * Fixed [#2042](https://github.com/rstudio/shiny/issues/2042), [#2628](https://github.com/rstudio/shiny/issues/2628): In a `dateInput` and `dateRangeInput`, disabled months and years are now a lighter gray, to make it easier to see that they are disabled. ([#2690](https://github.com/rstudio/shiny/pull/2690))
@@ -26,6 +28,8 @@ shiny 1.4.0.9001
 * `getDefaultReactiveDomain()` can now be called inside a `session$onSessionEnded` callback and will return the calling `session` information. ([#2757](https://github.com/rstudio/shiny/pull/2757))
 
 * Added a `'function'` class to `reactive()` and `reactiveVal()` objects. ([#2793](https://github.com/rstudio/shiny/pull/2793))
+
+* Added function `tabPanelBody()` which wraps `tabPanel()` without the title argument.  This function should be used within `tabsetPanel(type = "hidden")` ([#2814](https://github.com/rstudio/shiny/pull/2814))
 
 ### Bug fixes